Require $ before macro parameters, Macro overloads
Macros can have overloads and raw syntax now, and require $ before every parameter (with a type).

Any input that doesn't start with $ to denote a special token is taken as syntax, so now even the commas to separate parameters count as raw syntax, so you could actually write macros without any separator, or a custom one now.

In the future I want to be able to nest macros and have variadic ones like in rust, just to push this tiny language to its limits